ServicesUtilを完全なsingletonに変更
